+++ jabkit/src/main/java/org/jabref/toolkit/commands/Search.java
@@ -47,9 +47,8 @@ class Search implements Runnable {
     @Option(names = {"--query"}, description = "Search query", required = true)
     private String query;
 
-    // [impl->req~jabkit.cli.input-flag~1]
-    @Option(names = {"--input"}, converter = CygWinPathConverter.class, description = "Input BibTeX file", required = true)
-    private Path inputFile;
+    @Mixin
+    private InputOption inputOption = new InputOption();
 
     @Option(names = {"--output"}, converter = CygWinPathConverter.class, description = "Output file")
     private Path outputFile;
@@ -59,6 +58,7 @@ class Search implements Runnable {
 
     @Override
     public void run() {
+        Path inputFile = inputOption.getInputFile();
         Optional<ParserResult> parserResult = JabKit.importFile(
                 inputFile,
                 "bibtex",